qdrantpush/requirements.txt
qdrant-client
fastembed
langchain-text-splitters
python-dotenv